@@ -29,6 +29,8 @@
     "nav-frontend-lenker-style": "0.2.5",
     "nav-frontend-lukknapp": "1.0.1",
     "nav-frontend-lukknapp-style": "0.2.4",
+    "nav-frontend-modal": "^1.0.19",
+    "nav-frontend-modal-style": "^0.3.23",
     "nav-frontend-paneler": "^1.0.7",
     "nav-frontend-paneler-style": "^0.3.11",
     "nav-frontend-skjema": "^1.0.20",
@@ -49,6 +51,7 @@
     "react-dom": "16.3.2",
     "react-intl": "2.4.0",
     "react-maskedinput": "4.0.1",
+    "react-modal": "^3.6.1",
     "react-motion": "^0.5.2",
     "react-redux": "5.0.7",
     "react-redux-form-validation": "1.0.1",
@@ -75,12 +78,12 @@
     "start:tekster": "npm run build:tekster",
     "start": "npm-run-all -p start:*",
     "build": "npm run delete-static && npm run build:tekster && npm run build:ts",
-    "integrationtest-build": "cross-env REACT_APP_MOCK_BES=true npm run heroku-postbuild",
+    "integrationtest-build": "node scripts/preHeroku.js && npm install && cross-env REACT_APP_MOCK_BES=true cross-env REACT_APP_MOCK='true' npm run build && node scripts/postHeroku.js",
     "heroku-serve": "node index.js",
-    "heroku-postbuild": "node scripts/preHeroku.js && npm install && cross-env REACT_APP_MOCK='true' npm run build && node scripts/postHeroku.js"
+    "heroku-postbuild": "node scripts/preHeroku.js && npm install && cross-env REACT_APP_MOCK_ENDRE_REG_LOP=true cross-env REACT_APP_MOCK='true' npm run build && node scripts/postHeroku.js"
   },
   "devDependencies": {
-    "@navikt/create-react-app-typescript-less": "^1.0.0-beta.14",
+    "@navikt/create-react-app-typescript-less": "^1.0.0-beta.16",
     "@types/classnames": "2.2.3",
     "@types/fetch-mock": "5.12.2",
     "@types/jest": "22.2.3",
@@ -112,6 +115,6 @@
     "testcafe": "^0.23.0",
     "testcafe-browser-provider-browserstack": "^1.5.0",
     "tree-kill": "^1.2.0",
-    "typescript": "^2.8.3"
+    "typescript": "^3.0.3"
   }
 }
