@@ -1,27 +1,29 @@
 import * as React from 'react';
+import { FormattedMessage } from 'react-intl';
 import { Normaltekst } from 'nav-frontend-typografi';
 import Veilederpanel from 'nav-frontend-veilederpanel';
-import { FormattedMessage } from 'react-intl';
 
-const utropstegnSvg = require('../fullfor/utropstegn.svg');
+import utropstegnSvg from '../fullfor/utropstegn.svg';
+import './info-for-ikke-arbeidssoker-uten-oppfolging.less';
 
 class InfoForIkkeArbeidssokerUtenOppfolging extends React.Component {
-
     render() {
         return (
-            <Veilederpanel
-                type="plakat"
-                svg={<img
-                    src={utropstegnSvg}
-                    alt="Informasjon"
-                    className="nav-veilederpanel__illustrasjon"
-                />}
-                kompakt={true}
-            >
-                <Normaltekst>
-                    <FormattedMessage id="info-for-ikke-arbeidssoker-uten-oppfolging-innhold"/>
-                </Normaltekst>
-            </Veilederpanel>
+            <div className="info-for-ikke-arbeidssoker">
+                <Veilederpanel
+                    type="plakat"
+                    svg={<img
+                        src={utropstegnSvg}
+                        alt="Informasjon"
+                        className="nav-veilederpanel__illustrasjon"
+                    />}
+                    kompakt={true}
+                >
+                    <Normaltekst>
+                        <FormattedMessage id="info-for-ikke-arbeidssoker-uten-oppfolging-innhold"/>
+                    </Normaltekst>
+                </Veilederpanel>
+            </div>
         );
     }
 }
