@@ -1,11 +1,13 @@
 import * as React from 'react';
-import { Normaltekst, Systemtittel, Element } from 'nav-frontend-typografi';
 import { FormattedMessage } from 'react-intl';
+import { Normaltekst, Systemtittel, Element } from 'nav-frontend-typografi';
 import { VEIENTILARBEID_MED_DAGPENGER_URL, VEIENTILARBEID_URL } from '../../ducks/api';
 import AvsjekkBilde from './avsjekk-bilde';
 import { erIE } from '../../utils/ie-test';
+import { frontendLogger } from '../../metrikker/metrics-utils';
 
-const handinfoSvg = require('./clipboard.svg');
+import handinfoSvg from './clipboard.svg';
+import './registrert.less';
 
 class DuErNaRegistrert extends React.Component {
 
@@ -33,10 +35,22 @@ class DuErNaRegistrert extends React.Component {
                             <FormattedMessage id="duernaregistrert-element"/>
                         </Element>
                         <div className="registrert__knapperad">
-                            <a href={VEIENTILARBEID_URL} className="registrert__lenke knapp knapp--standard">
+                            <a
+                                href={VEIENTILARBEID_URL}
+                                className="registrert__lenke knapp knapp--standard"
+                                onClick={() => {
+                                    frontendLogger('registrering.ikke.vis.dagpenger.info');
+                                }}
+                            >
                                 <FormattedMessage id="knapp-ikke-na"/>
                             </a>
-                            <a href={VEIENTILARBEID_MED_DAGPENGER_URL} className="registrert__lenke knapp knapp--hoved">
+                            <a
+                                href={VEIENTILARBEID_MED_DAGPENGER_URL}
+                                className="registrert__lenke knapp knapp--hoved"
+                                onClick={() => {
+                                    frontendLogger('registrering.vis.dagpenger.info');
+                                }}
+                            >
                                 <FormattedMessage id="knapp-ja-vis-meg"/>
                             </a>
                         </div>
