@@ -11,16 +11,19 @@ import {
     basename,
 } from './utils/konstanter';
 import Routes from './routes';
+import Modal from 'react-modal';
 import HerokuappEndreMockRegistreringLoep from
     './mocks/HerokuappEndreMockRegistreringLoep/herokuapp-endre-mock-registrering-loep';
 
 const store = getStore();
 
+Modal.setAppElement('#root');
+
 class App extends React.Component {
 
     render() {
         const BrukesIHerokuappEndreRegistreringTypeMock =
-            !process.env.REACT_APP_MOCK_BES
+            process.env.REACT_APP_MOCK_ENDRE_REG_LOP
             ? <HerokuappEndreMockRegistreringLoep/>
             : null;
         return (
