@@ -1,12 +1,10 @@
 import { fetchToJson, fetchWithTimeout } from './api-utils';
-import {Data as RegistrerBrukerData } from './registrerbruker';
+import { Data as RegistrerBrukerData } from './registrerbruker';
 import { alleFeatureToggles } from './feature-toggles';
 
 export const INNLOGGINGSLINJE_URL = '/innloggingslinje/auth';
 export const AUTENTISERINGSINFO_URL = '/veilarbstepup/status';
 export const SBLARBEID_URL = '/sbl/nav_security_check?goto=/sbl/arbeid/endreCv';
-export const DITTNAV_URL = '/dittnav/';
-export const FORSIDENAV_URL = 'https://www.nav.no/';
 export const VEIENTILARBEID_URL = '/veientilarbeid/';
 export const VEIENTILARBEID_MED_DAGPENGER_URL = '/veientilarbeid/?visInformasjonsmodul=true&visdagpenger=true';
 export const ARBEIDSSOKERREGISTRERING_START = '/arbeidssokerregistrering/start';
@@ -14,6 +12,8 @@ export const VEILARBSTEPUP = `/veilarbstepup/oidc?url=${ARBEIDSSOKERREGISTRERING
 export const SBLARBEID_OPPRETT_MIN_ID_URL = '/sbl/nav_security_check?goto=/sbl/arbeid/opprettMinIdBruker';
 export const VEILARBREGISTRERING_URL = '/veilarbregistrering/api';
 export const FEATURE_URL = '/feature';
+export const DITT_NAV_URL = '/dittnav';
+export const DITT_SYKEFRAVAER_URL = '/sykefravaer';
 
 const PAM_JANZZ_URL = '/pam-janzz/rest';
 const STYRK_URL = `${PAM_JANZZ_URL}/typeahead/yrke-med-styrk08`;
@@ -38,7 +38,8 @@ const MED_CREDENTIALS = {
 export function hentRegistreringStatus() {
     return fetchToJson({
         url: `${VEILARBREGISTRERING_URL}/startregistrering`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         }
     });
@@ -47,20 +48,24 @@ export function hentRegistreringStatus() {
 export function registrerBruker(data: RegistrerBrukerData) {
     return fetchToJson({
         url: `${VEILARBREGISTRERING_URL}/startregistrering`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
             method: 'post',
-            body: JSON.stringify(data)}
+            body: JSON.stringify(data)
+        }
     });
 }
 
 export function startReaktivering() {
     return fetchToJson({
         url: `${VEILARBREGISTRERING_URL}/startreaktivering`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
             method: 'post',
-            body: JSON.stringify({})}
+            body: JSON.stringify({})
+        }
     });
 }
 
@@ -80,7 +85,8 @@ export function registrerBrukerSBLArbeid(timeoutMillis?: number) {
 export function hentBrukersNavn() {
     return fetchToJson({
         url: `${INNLOGGINGSLINJE_URL}?randomness=${Math.random()}`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         }
     });
@@ -89,7 +95,8 @@ export function hentBrukersNavn() {
 export function hentAutentiseringsInfo() {
     return fetchToJson({
         url: `${AUTENTISERINGSINFO_URL}`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         }
     });
@@ -98,7 +105,8 @@ export function hentAutentiseringsInfo() {
 export function hentStyrkkodeForSisteStillingFraAAReg() {
     return fetchToJson({
         url: `${VEILARBREGISTRERING_URL}/sistearbeidsforhold`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         },
         recoverWith: () => ({arbeidsgiver: null, stilling: null, styrk: null, fra: null, til: null})
@@ -108,7 +116,8 @@ export function hentStyrkkodeForSisteStillingFraAAReg() {
 export function hentStillingFraPamGittStyrkkode(styrk: string) {
     return fetchToJson({
         url: `${PAM_JANZZ_URL}/kryssklassifiserMedKonsept?kodeForOversetting=${styrk}`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         },
         recoverWith: () => ({konseptMedStyrk08List: []})
@@ -118,7 +127,8 @@ export function hentStillingFraPamGittStyrkkode(styrk: string) {
 export function hentStillingMedStyrk08(sokestreng: string) {
     return fetchToJson({
         url: `${STYRK_URL}?q=${sokestreng}`,
-        config: {...{redirect: 'manual'},
+        config: {
+            ...{redirect: 'manual'},
             headers: getHeaders()
         },
         recoverWith: () => ({'typeaheadYrkeList': []})
@@ -129,7 +139,8 @@ export function hentFeatureToggles() {
     const parameters = alleFeatureToggles.map(element => 'feature=' + element).join('&');
     return fetchToJson({
         url: `${FEATURE_URL}/?${parameters}`,
-        config: { ...MED_CREDENTIALS,
+        config: {
+            ...MED_CREDENTIALS,
             headers: getHeaders(),
         },
         recoverWith: () => ({})
