@@ -1,8 +1,10 @@
 import * as React from 'react';
 import { FormattedHTMLMessage } from 'react-intl';
 import { Normaltekst } from 'nav-frontend-typografi';
-import { VEILARBSTEPUP } from '../../ducks/api';
 import NavAlertStripe from 'nav-frontend-alertstriper';
+import { VEILARBSTEPUP } from '../../ducks/api';
+
+import './stepup.less';
 
 function StepUp() {
     return (
@@ -13,7 +15,7 @@ function StepUp() {
             <div className="knapperad">
                 <button
                     className="knapp knapp--hoved stepup__knapp"
-                    onClick={() => document.location.href = VEILARBSTEPUP}
+                    onClick={() => window.location.href = VEILARBSTEPUP}
                 >
                     <Normaltekst>
                         <FormattedHTMLMessage id="knapp-logg-inn"/>
