@@ -2,14 +2,16 @@ import * as React from 'react';
 import { Innholdstittel, Normaltekst } from 'nav-frontend-typografi';
 import { scrollToBanner } from '../../utils/utils';
 
+import './loader.less';
+
 export const loaderTittelElement = (
     <>
-    <Innholdstittel className="blokk-s">
-        Registrering pågår
-    </Innholdstittel>
-    <Normaltekst>
-        Vi setter opp tjenester til deg. Dette kan ta noen sekunder.
-    </Normaltekst>
+        <Innholdstittel className="blokk-s">
+            Registrering pågår
+        </Innholdstittel>
+        <Normaltekst>
+            Vi setter opp tjenester til deg. Dette kan ta noen sekunder.
+        </Normaltekst>
     </>
 );
 
