@@ -16,15 +16,15 @@ class ProgressBarContainer extends React.Component<Props> {
     render() {
         // Tar med :url fra Router for å oppdatere progressbar når url endres
 
-        const pathName = document.location.pathname.replace(basename, '');
-        const config = finnRiktigConfig(pathName);
+        const pathname = this.props.location.pathname.toString().replace(basename, '');
+        const config = finnRiktigConfig(pathname);
 
         if (config.length === 0) {
             return null;
         } else {
             return (
                 <ProgressBar
-                    gjeldendeSporsmal={config.indexOf(pathName)}
+                    gjeldendeSporsmal={config.indexOf(pathname)}
                     antallSporsmal={config.length}
                     offset={3}
                 />
