@@ -15,9 +15,8 @@ import Routes from './routes';
 import { RegistreringType } from './ducks/registreringstatus';
 import Fullfor from './sider/fullfor/fullfor';
 import {
-    AVBRYT_PATH,
     DU_ER_NA_REGISTRERT_PATH,
-    FULLFOR_PATH,
+    FULLFOR_PATH, INNGANGSSPORSMAL_PATH,
     OPPSUMMERING_PATH,
     REAKTIVERING_PATH,
     START_PATH
@@ -33,8 +32,8 @@ import {Store} from 'redux';
 import {AppState} from './reducer';
 import DuErNaRegistrert from './sider/registrert/registrert';
 import { ActionTypes as ReaktiverBrukerActionTypes } from './ducks/reaktiverbruker';
-import Avbryt from './sider/avbryt/avbryt';
 import KreverReaktivering from './sider/krever-reaktivering/krever-reaktivering';
+import Startside from './sider/start/startside';
 
 enzyme.configure({adapter: new Adapter()});
 
@@ -124,7 +123,6 @@ describe('Routes', () => {
 
     });
 
-
     it('Skal gå til første side istedenfor fullfør hvis spørsmål ikke er besvart', () => {
 
         const store = create();
@@ -142,7 +140,6 @@ describe('Routes', () => {
         expect(wrapper.find({ to: START_PATH })).to.have.length(1);
 
     });
-
 
     it('Skal gå til oppsummering side hvis alle spørsmål er besvart', () => {
 
@@ -162,7 +159,6 @@ describe('Routes', () => {
 
     });
 
-
     it('Skal gå til første side istedenfor oppsummering hvis spørsmål ikke er besvart', () => {
 
         const store = create();
@@ -176,7 +172,6 @@ describe('Routes', () => {
         dispatchNoenSporsmal(store);
 
         const wrapper = mountWithStoreRouterAndIntl(<Routes />, store, [OPPSUMMERING_PATH]);
-
         expect(wrapper.find({ to: START_PATH })).to.have.length(1);
 
     });
@@ -233,27 +228,10 @@ describe('Routes', () => {
 
     });
 
-
-    it('Skal gå til avbryt side', () => {
+    it('Skal sende bruker til sbl om den ikke er under oppfølging og ny-registrering er avskrudd', () => {
 
         const store = create();
-
         dispatchRegistreringstatus({ registreringType: RegistreringType.ORDINAER_REGISTRERING }, store);
-        dispatchFeaturestatus({
-            'arbeidssokerregistrering.gradual-rollout-ny-registrering': true,
-            'arbeidssokerregistrering.sykefravaer': true
-        }, store);
-
-        const wrapper = mountWithStoreRouterAndIntl(<Routes />, store, [AVBRYT_PATH]);
-
-        expect(wrapper.find(Avbryt)).to.have.length(1);
-
-    });
-
-
-    it('skal sende bruker til sbl om den ikke er under oppfølging og ny-registrering er avskrudd', () => {
-
-        const store = create();
 
         dispatchFeaturestatus({
             'arbeidssokerregistrering.gradual-rollout-ny-registrering': false,
@@ -261,12 +239,11 @@ describe('Routes', () => {
         }, store);
 
         const wrapper = mountWithStoreRouterAndIntl(<Routes />, store);
-
         expect(wrapper.find(SblRegistrering)).to.have.length(1);
 
     });
 
-    it('skal ikke sende bruker til sbl om feature for ny-registrering er påskrudd', () => {
+    it('Skal ikke sende bruker til sbl om feature for ny-registrering er påskrudd', () => {
         const store = create();
         dispatchFeaturestatus({
             'arbeidssokerregistrering.gradual-rollout-ny-registrering': true,
@@ -279,8 +256,7 @@ describe('Routes', () => {
 
     });
 
-
-    it('skal sende bruker til AlleredeRegistrert om den er under oppfølging', () => {
+    it('Skal sende bruker til AlleredeRegistrert om den er under oppfølging', () => {
         const store = create();
 
         dispatchRegistreringstatus({ registreringType: RegistreringType.ALLEREDE_REGISTRERT }, store);
@@ -290,7 +266,7 @@ describe('Routes', () => {
         expect(wrapper.find(AlleredeRegistrert)).to.have.length(1);
     });
 
-    it('skal sende bruker til InfoForIkkeArbeidssokerUtenOppfolging om status er erIkkeArbeidssokerUtenOppfolging', () => {
+    it('Skal sende bruker til InfoForIkkeArbeidssokerUtenOppfolging når status er sperret', () => {
         const store = create();
 
         dispatchRegistreringstatus({ registreringType: RegistreringType.SPERRET }, store);
@@ -315,5 +291,52 @@ describe('Routes', () => {
         expect(wrapper.find({ to: START_PATH })).to.have.length(1);
 
     });
+
+    it('Skal redirecte til KreverReaktivering hvis registreringstype er REAKTIVERING', () => {
+
+        const store = create();
+
+        dispatchRegistreringstatus({ registreringType: RegistreringType.REAKTIVERING }, store);
+        dispatchFeaturestatus({
+            'arbeidssokerregistrering.gradual-rollout-ny-registrering': true,
+            'arbeidssokerregistrering.sykefravaer': true
+        }, store);
+
+        const wrapper = mountWithStoreRouterAndIntl(<Routes />, store, [START_PATH]);
+
+        expect(wrapper.find(KreverReaktivering)).to.have.length(1);
+
+    });
+
+    it('Skal redirecte til Inngangssporsmal dersom ny registrering er på', () => {
+
+        const store = create();
+
+        dispatchRegistreringstatus({registreringType: RegistreringType.SYKMELDT_REGISTRERING, sykmeldtFraDato: "2018-10-21"}, store);
+        dispatchFeaturestatus({
+            'arbeidssokerregistrering.gradual-rollout-ny-registrering': true,
+            'arbeidssokerregistrering.sykefravaer': true
+        }, store);
+
+        const wrapper = mountWithStoreRouterAndIntl(<Routes />, store, [START_PATH]);
+        expect(wrapper.find({ to: INNGANGSSPORSMAL_PATH })).to.have.length(1);
+
+    });
+
+    it('Skal gå til start dersom ny registrering ikke er på', () => {
+
+        const store = create();
+
+        dispatchRegistreringstatus({registreringType: RegistreringType.SYKMELDT_REGISTRERING}, store);
+        dispatchFeaturestatus({
+            'arbeidssokerregistrering.gradual-rollout-ny-registrering': true,
+            'arbeidssokerregistrering.sykefravaer': false
+        }, store);
+
+        const wrapper = mountWithStoreRouterAndIntl(<Routes />, store, [START_PATH]);
+        expect(wrapper.find(Startside)).to.have.length(1);
+
+    });
+
 
 });
