@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
+import { FormattedMessage } from 'react-intl';
+import { RouteComponentProps } from 'react-router';
 import KnappBase from 'nav-frontend-knapper';
 import { Innholdstittel, Normaltekst } from 'nav-frontend-typografi';
-import { FormattedMessage } from 'react-intl';
 import { disableVerikalScrollingVedAnimasjon, MatchProps } from '../../utils/utils';
-import { RouteComponentProps } from 'react-router';
 import { AppState } from '../../reducer';
 import { FULLFOR_PATH, START_PATH } from '../../utils/konstanter';
 import LenkeAvbryt from '../../komponenter/knapper/lenke-avbryt';
@@ -15,6 +15,8 @@ import OrdinaerOppsummeringBesvarelser from './ordinaer-oppsummering-besvarelser
 import SykmeldtOppsummeringBesvarelser from './sykmeldt-oppsummering-besvarelser';
 import { erKlarForFullforing } from '../fullfor/fullfor-utils';
 
+import './oppsummering.less';
+
 interface StateProps {
     state: AppState;
 }
@@ -24,16 +26,14 @@ type Props = StateProps & RouteComponentProps<MatchProps>;
 class Oppsummering extends React.Component<Props> {
 
     componentWillMount() {
-
         if (!erKlarForFullforing(this.props.state)) {
             this.props.history.push(START_PATH);
         }
-      
         disableVerikalScrollingVedAnimasjon();
     }
 
     render() {
-        const { history, state } = this.props;
+        const {history, state} = this.props;
         let classnames = 'oppsummering ';
         classnames += erIE() ? 'erIE' : '';
 
@@ -49,21 +49,19 @@ class Oppsummering extends React.Component<Props> {
         return (
             <section className={classnames}>
                 <Innholdstittel tag="h1" className="oppsummering-tittel">
-                    <FormattedMessage id={`${tekstPrefix}-tittel`} />
+                    <FormattedMessage id={`${tekstPrefix}-tittel`}/>
                 </Innholdstittel>
                 <Normaltekst className="oppsummering-ingress">
                     <FormattedMessage id={`${tekstPrefix}-ingress`}/>
                 </Normaltekst>
-
                 {oppsummeringBesvarelser}
-
-                <div className="knapper-vertikalt">
+                <div className="lenke-avbryt-wrapper">
                     <KnappBase type="hoved" onClick={() => history.push(FULLFOR_PATH)} data-testid="neste">
                         <FormattedMessage id="knapp-riktig"/>
                     </KnappBase>
-                    <LenkeTilbake onClick={() => this.props.history.goBack()}/>
-                    <LenkeAvbryt wrapperClassname="no-anim"/>
                 </div>
+                <LenkeTilbake onClick={() => this.props.history.goBack()}/>
+                <LenkeAvbryt wrapperClassname="wrapper-too"/>
             </section>
         );
     }
