@@ -16,7 +16,9 @@ function UtdanningGodkjentSporsmal(props: Props) {
         getTekstId: (svar: Svar) => getTekstIdForSvar(props.sporsmalId, svar),
         hentAvgittSvar: () => props.hentAvgittSvar(props.sporsmalId),
     };
-    const getTekst = (kontekst: TekstKontekst) => getIntlTekstForSporsmal(props.sporsmalId, kontekst, props.intl);
+    const getTekst = (kontekst: TekstKontekst) => getIntlTekstForSporsmal(props.sporsmalId,
+        kontekst, props.intl, props.registeringType);
+
     return (
         <form className="spm-skjema">
             <fieldset className="skjema__fieldset">
