@@ -14,6 +14,7 @@ import Utdanningsporsmal from '../skjema-registrering/sporsmal/sporsmal-utdannin
 import UtdanningGodkjentSporsmal from '../skjema-registrering/sporsmal/sporsmal-utdanning-godkjent';
 import UtdanningBestattSporsmal from '../skjema-registrering/sporsmal/sporsmal-utdanning-bestatt';
 import AndreForhold from '../skjema-registrering/sporsmal/sporsmal-andre-forhold';
+import { RegistreringType } from '../../ducks/registreringstatus';
 
 interface DispatchProps {
     endreSvar: (sporsmalId: string, svar: Svar) => void;
@@ -40,6 +41,8 @@ class SkjemaSykefravaerUsikker extends React.Component<Props> {
             hentAvgittSvar: (sporsmalId: SporsmalId) => hentSvar(svarState, sporsmalId),
         };
 
+        const regType = RegistreringType.SYKMELDT_REGISTRERING;
+
         return (
             <Skjema
                 config={skjemaFlytUsikker}
@@ -47,10 +50,26 @@ class SkjemaSykefravaerUsikker extends React.Component<Props> {
                 endUrl={OPPSUMMERING_PATH}
                 {...{location, match, history}}
             >
-                <Utdanningsporsmal sporsmalId={SporsmalId.utdanning} {...fellesProps}/>
-                <UtdanningGodkjentSporsmal sporsmalId={SporsmalId.utdanningGodkjent} {...fellesProps}/>
-                <UtdanningBestattSporsmal sporsmalId={SporsmalId.utdanningBestatt} {...fellesProps}/>
-                <AndreForhold sporsmalId={SporsmalId.andreForhold} {...fellesProps}/>
+                <Utdanningsporsmal
+                    sporsmalId={SporsmalId.utdanning}
+                    {...fellesProps}
+                    registeringType={regType}
+                />
+                <UtdanningGodkjentSporsmal
+                    sporsmalId={SporsmalId.utdanningGodkjent}
+                    {...fellesProps}
+                    registeringType={regType}
+                />
+                <UtdanningBestattSporsmal
+                    sporsmalId={SporsmalId.utdanningBestatt}
+                    {...fellesProps}
+                    registeringType={regType}
+                />
+                <AndreForhold
+                    sporsmalId={SporsmalId.andreForhold}
+                    {...fellesProps}
+                    registeringType={regType}
+                />
             </Skjema>
         );
     }
