@@ -5,12 +5,11 @@ export const SKJEMA_PATH = '/skjema';
 export const SKJEMA_SYKEFRAVAER_PATH = '/skjema-sykefravaer';
 export const OPPSUMMERING_PATH = '/oppsummering';
 export const SBLREG_PATH = '/sblregistrering';
-export const AVBRYT_PATH = '/avbryt';
 export const FULLFOR_PATH = '/fullfor';
 export const DU_ER_NA_REGISTRERT_PATH = '/duernaregistrert';
 export const ALLEREDE_REGISTRERT_PATH = '/allerederegistrert';
 export const REAKTIVERING_PATH = '/reaktivering';
-export const IKKE_ARBEIDSSSOKER_UTENFOR_OPPFOLGING = '/ikke-arbeidssoker-utenfor-oppfolging';
+export const IKKE_ARBEIDSSSOKER_UTENFOR_OPPFOLGING_PATH = '/ikke-arbeidssoker-utenfor-oppfolging';
 
 /* Besvarelser */
 
